django/requirements.txt
appnope==0.1.0
coverage==3.7.1
decorator==4.0.2
defusedxml==0.4.1
Django==1.8.4
django-cors-headers==1.1.0
django-debug-toolbar==1.3.2
django-extensions==1.5.6
django-filter==0.10.0
django-grappelli==2.7.1
djangorestframework==3.2.3
factory-boy==2.5.2
gnureadline==6.3.3
ipython==4.0.0
ipython-genutils==0.1.0
Markdown==2.6.2
oauthlib==1.0.1
path.py==8.1
pexpect==3.3
pickleshare==0.5
psycopg2==2.6.1
python3-openid==3.0.6
requests==2.7.0
requests-oauthlib==0.5.0
rules==0.4
simplegeneric==0.8.1
six==1.9.0
sqlparse==0.1.16
traitlets==4.0.0
wheel==0.24.0
